feat(front): 메인 페이지(장바구니 버전) 구현
