+++ coffee-front/src/app/admin/page.tsx
@@ -1,13 +1,24 @@
 "use client";
 
 import { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
 import Link from "next/link";
 import { fetchApi } from "@/lib/client";
 import type { ProductDto } from "@/type/product";
 
 export default function ProductListPage() {
+  const router = useRouter();
   const [items, setItems] = useState<ProductDto[] | null>(null);
   const [err, setErr] = useState<string | null>(null);
+
+  // 🔒 관리자 접근 제어
+  useEffect(() => {
+    const role = parseInt(localStorage.getItem("role") || "0");
+    if (role !== 1) {
+      alert("관리자 권한이 필요합니다.");
+      router.back; // 일반 페이지로 이동
+    }
+  }, [router]);
 
   const load = () => {
     fetchApi<ProductDto[]>("/api/products")
@@ -29,6 +40,17 @@ export default function ProductListPage() {
 
   return (
     <main className="p-4 max-w-3xl mx-auto">
+      <header className="flex justify-end gap-3 mb-6">
+        <button
+          className="px-4 py-2 border rounded hover:bg-gray-50"
+          onClick={() => {
+            router.push("/logout");
+          }}
+        >
+          로그아웃
+        </button>
+      </header>
+
       <div className="flex items-center justify-between mb-4">
         <h1 className="text-2xl font-bold">상품 관리</h1>
         <Link href="/admin/new" className="border rounded px-3 py-2">추가</Link>
@@ -39,39 +61,35 @@ export default function ProductListPage() {
       {items?.length === 0 && <div>상품이 없습니다.</div>}
 
       {items && items.length > 0 && (
-  <ul className="divide-y">
-    {items.map(p => (
-      <li key={p.id} className="py-3 flex items-center justify-between">
-        <div className="flex items-center gap-3">
-          {/* 썸네일 */}
-          <img
-            src={p.imageUrl || "/images/placeholder-product.png"}
-            alt={p.name}
-            loading="lazy"
-            className="w-20 h-20 object-cover rounded border"
-            onError={(e) => {
-              e.currentTarget.src = "/images/placeholder-product.png";
-            }}
-          />
+        <ul className="divide-y">
+          {items.map(p => (
+            <li key={p.id} className="py-3 flex items-center justify-between">
+              <div className="flex items-center gap-3">
+                <img
+                  src={p.imageUrl || "/images/placeholder-product.png"}
+                  alt={p.name}
+                  loading="lazy"
+                  className="w-20 h-20 object-cover rounded border"
+                  onError={(e) => {
+                    e.currentTarget.src = "/images/placeholder-product.png";
+                  }}
+                />
+                <div>
+                  <div className="font-semibold">{p.name}</div>
+                  <div className="text-sm text-gray-600">
+                    ₩{p.price.toLocaleString()} · 재고 {p.stock}
+                  </div>
+                </div>
+              </div>
 
-          {/* 텍스트 */}
-          <div>
-            <div className="font-semibold">{p.name}</div>
-            <div className="text-sm text-gray-600">
-              ₩{p.price.toLocaleString()} · 재고 {p.stock}
-            </div>
-          </div>
-        </div>
-
-        <div className="flex gap-2">
-          <Link href={`/admin/${p.id}/edit`} className="border rounded px-2 py-1">수정</Link>
-          <button onClick={() => onDelete(p.id)} className="border rounded px-2 py-1">삭제</button>
-        </div>
-      </li>
-    ))}
-  </ul>
-)}
-
+              <div className="flex gap-2">
+                <Link href={`/admin/${p.id}/edit`} className="border rounded px-2 py-1">수정</Link>
+                <button onClick={() => onDelete(p.id)} className="border rounded px-2 py-1">삭제</button>
+              </div>
+            </li>
+          ))}
+        </ul>
+      )}
     </main>
   );
 }
